% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{List available palette names}
\usage{
list_palettes()
}
\value{
A character vector of palette names.
}
\description{
List available palette names
}
\examples{
list_palettes()
}
